air_pollution_project/requirements.txt
flask
torch
torchvision
pillow
opencv-python
numpy
werkzeug
